Video player/DLNA player redesign

+++ shared/src/commonMain/kotlin/com/ismartcoding/plain/ui/page/chat/components/ChatAudioInlineControls.kt
@@ -32,7 +32,7 @@ fun ChatAudioInlineControls(
     duration: Float,
     isPlaying: Boolean,
     onProgressChange: (Float) -> Unit,
-    onValueChangeFinished: () -> Unit,
+    onValueChangeFinished: (Float) -> Unit,
     onShowFullPlayer: () -> Unit,
     onPause: () -> Unit,
     onPlay: () -> Unit,
